+++ js/route.js
@@ -1,0 +1,13 @@
+import { loginRender, joinRender } from "./main"
+
+export function router() {
+    const routePath = location.hash
+
+    if (routePath.includes('#login')) {
+        loginRender()
+    }
+    else if (routePath.includes('#join')) {
+        joinRender()
+    }
+}
+
