+++ js/test.js
@@ -1,17 +1,16 @@
-import {getLogOut} from './getdata.js'
+import { getLogOut } from './getdata.js'
+import { logOut } from './main.js'
 
-export function completeLogin(displayName){
+/* 로그인 완료시 나오는  */
+
+export function completeLogin(displayName) {
   const li = document.querySelector('.signUpsignIn')
   li.innerHTML = /*HTML*/ `
     <p><strong>${displayName}</strong>님, 환영합니다.</p>
-    <a class="logOutBtn" href="#">로그아웃</a>
+    <a class="logOutBtn">로그아웃</a>
     `
-
-    const logOutBtn = li.querySelector('.logOutBtn')
-    logOutBtn.addEventListener('click', async () => {
-      const accessToken = localStorage.getItem('accessToken')
-      console.log(accessToken)
-      const res = await getLogOut(accessToken)
-      console.log(res)
-    })
+  logOut()
 }
+
+/* 로그아웃 버튼 눌렀을때 토큰 삭제하고 창 없애기 */
+
